pyproject.toml
[project]
name = "twilio-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "twilio>=9.6.3",
]
